Packaging + Launcher + Bootstrap
- Ship pre-built core/dashboard jars in build/ from package-release.sh
- Extend iw-run with 3-tier jar resolution (env -> build/ -> Mill)
- Consolidate review-state command into single file (drops runtime core/ walk)
